+++ tests/gunicorn/files/gconf.py
@@ -1,1 +1,1 @@
-worker_class = 'async'
+worker_class = 'gthread'